+++ src/Either.ts
@@ -15,6 +15,7 @@
  */
 import { Alt2, Alt2C } from './Alt'
 import { Applicative as ApplicativeHKT, Applicative2, Applicative2C } from './Applicative'
+import { Apply2, apS_ } from './Apply'
 import { Bifunctor2 } from './Bifunctor'
 import { ChainRec2, ChainRec2C, tailRec } from './ChainRec'
 import { Separated } from './Compactable'
@@ -1147,6 +1148,16 @@ export const Functor: Functor2<URI> = {
 
 /**
  * @category instances
+ * @since 2.10.0
+ */
+export const Apply: Apply2<URI> = {
+  URI,
+  map: _map,
+  ap: _ap
+}
+
+/**
+ * @category instances
  * @since 2.7.0
  */
 export const Applicative: Applicative2<URI> = {
@@ -1374,22 +1385,17 @@ export const bind: <N extends string, A, E, B>(
 /**
  * @since 2.8.0
  */
-export const apSW = <A, N extends string, D, B>(
-  name: Exclude<N, keyof A>,
-  fb: Either<D, B>
-): (<E>(fa: Either<E, A>) => Either<D | E, { [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
-  flow(
-    map((a) => (b: B) => bind_(a, name, b)),
-    apW(fb)
-  )
+export const apS =
+  /*#__PURE__*/
+  apS_(Apply)
 
 /**
  * @since 2.8.0
  */
-export const apS: <A, N extends string, E, B>(
+export const apSW: <A, N extends string, D, B>(
   name: Exclude<N, keyof A>,
-  fb: Either<E, B>
-) => (fa: Either<E, A>) => Either<E, { [K in keyof A | N]: K extends keyof A ? A[K] : B }> = apSW
+  fb: Either<D, B>
+) => <E>(fa: Either<E, A>) => Either<D | E, { [K in keyof A | N]: K extends keyof A ? A[K] : B }> = apS as any
 
 // -------------------------------------------------------------------------------------
 // array utils