+++ src/Task.ts
@@ -11,6 +11,7 @@
  * @since 2.0.0
  */
 import { Applicative1 } from './Applicative'
+import { Apply1, apS_ } from './Apply'
 import { bind_, flow, identity, pipe } from './function'
 import { bindTo_, Functor1 } from './Functor'
 import { IO } from './IO'
@@ -308,6 +309,16 @@ export const Functor: Functor1<URI> = {
 
 /**
  * @category instances
+ * @since 2.10.0
+ */
+export const ApplyPar: Apply1<URI> = {
+  URI,
+  map: _map,
+  ap: _apPar
+}
+
+/**
+ * @category instances
  * @since 2.7.0
  */
 export const ApplicativePar: Applicative1<URI> = {
@@ -315,6 +326,16 @@ export const ApplicativePar: Applicative1<URI> = {
   map: _map,
   ap: _apPar,
   of
+}
+
+/**
+ * @category instances
+ * @since 2.10.0
+ */
+export const ApplySeq: Apply1<URI> = {
+  URI,
+  map: _map,
+  ap: _apSeq
 }
 
 /**
@@ -423,14 +444,9 @@ export const bind = <N extends string, A, B>(
 /**
  * @since 2.8.0
  */
-export const apS = <A, N extends string, B>(
-  name: Exclude<N, keyof A>,
-  fb: Task<B>
-): ((fa: Task<A>) => Task<{ [K in keyof A | N]: K extends keyof A ? A[K] : B }>) =>
-  flow(
-    map((a) => (b: B) => bind_(a, name, b)),
-    ap(fb)
-  )
+export const apS =
+  /*#__PURE__*/
+  apS_(ApplyPar)
 
 // -------------------------------------------------------------------------------------
 // array utils